requirements.txt
fastapi
uvicorn
httpx
python-dotenv
google-generativeai
beautifulsoup4
lxml